# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CS2 Server Management Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface CreateUserBody {
  username: string;
  password: string;
  role: CreateUserBodyRole;
}

export interface Server {
  id: number;
  name: string;
  ip: string;
  port: number;
  agentUrl: string;
  agentToken: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateServerBody {
  name: string;
  ip: string;
  port: number;
  agentUrl: string;
  agentToken: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateServerBody {
  name?: string;
  ip?: string;
  port?: number;
  agentUrl?: string;
  agentToken?: string;
  /** @nullable */
  description?: string | null;
}

export interface ServerStatus {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  /** @nullable */
  map?: string | null;
  /** @nullable */
  cpuUsage?: number | null;
  /** @nullable */
  ramUsage?: number | null;
  /** @nullable */
  uptime?: string | null;
  /** @nullable */
  tickrate?: number | null;
  agentReachable: boolean;
}

/**
 * @nullable
 */
export type AgentResponseData = { [key: string]: unknown } | null;

export interface AgentResponse {
  success: boolean;
  message: string;
  /** @nullable */
  data?: AgentResponseData;
}

export interface RconCommandBody {
  command: string;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  debug: "debug",
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
}

export interface Player {
  steamId: string;
  name: string;
  score: number;
  ping: number;
  duration: string;
  /** @nullable */
  ip?: string | null;
}

export interface PlayerActionBody {
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  duration?: number | null;
}

export interface Plugin {
  id: string;
  name: string;
  version: string;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  description?: string | null;
  enabled: boolean;
}

export interface DashboardSummary {
  totalServers: number;
  onlineServers: number;
  totalPlayers: number;
  totalUsers: number;
}

export type ActivityEntryType =
  (typeof ActivityEntryType)[keyof typeof ActivityEntryType];

export const ActivityEntryType = {
  server_start: "server_start",
  server_stop: "server_stop",
  server_restart: "server_restart",
  player_kick: "player_kick",
  player_ban: "player_ban",
  plugin_change: "plugin_change",
  user_login: "user_login",
  rcon_command: "rcon_command",
} as const;

export interface ActivityEntry {
  id: number;
  type: ActivityEntryType;
  /** @nullable */
  serverId?: number | null;
  /** @nullable */
  serverName?: string | null;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userName?: string | null;
  details: string;
  createdAt: string;
}

export type GetServerLogsParams = {
  lines?: number;
};